feat(logfx): implement debug, err log
