+++ app/gqlclient/log.go
@@ -6,6 +6,10 @@ func (u *UseCase) AppInfo(ctx context.Context, format string, a ...any) {
 	u.li.Log.AppInfo(ctx, format, a...)
 }
 
-func (u *UseCase) AppDebugE(ctx context.Context, err error) {
-	//
+func (u *UseCase) AppError(ctx context.Context, err error) {
+	u.li.Log.Error(ctx, err)
+}
+
+func (u *UseCase) AppDebug(ctx context.Context, format string, a... any) {
+	u.li.Log.Debug(ctx, format, a...)
 }
