+++ lib/logfx/main.go
@@ -10,7 +10,9 @@ func New(repository IRepository) ILogSrv {
 
 type ILogSrv interface {
 	AppInfo(ctx context.Context, format string, a ...any)
+	Error(ctx context.Context, err error)
 	Info(ctx context.Context, format string, a ...any)
+	Debug(ctx context.Context, format string, a ...any)
 	Faint(ctx context.Context, format string, a ...any)
 }
